% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athlete_activities.R
\name{athlete_activities}
\alias{athlete_activities}
\title{List Athlete Activities}
\usage{
athlete_activities(
  client = strava_client(),
  before = NULL,
  after = NULL,
  page = 1,
  per_page = 30,
  ...
)
}
\arguments{
\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}

\item{before}{An epoch timestamp to use for filtering activities that have
taken place before a certain time.}

\item{after}{An epoch timestamp to use for filtering activities that have
taken place after a certain time.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{...}{Additional parameters for \code{httr2::req_oauth_auth_code}.}
}
\description{
Returns the activities of an athlete for a specific identifier. Requires
activity:read. Only Me activities will be filtered out unless requested by a
token with activity:read_all.
}
\examples{
if (FALSE) {
  athlete_activities()
}

}
