% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities.R
\name{get_activities}
\alias{get_activities}
\title{Get Activity}
\usage{
get_activities(
  id,
  include_all_efforts = NULL,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the activity.}

\item{include_all_efforts}{To include all segments efforts.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\value{
The activity's detailed representation. An instance of
\href{https://developers.strava.com/docs/reference/#api-models-DetailedActivity}{DetailedActivity}.
}
\description{
Returns the given activity that is owned by the authenticated athlete.
Requires \code{activity:read} for Everyone and Followers activities. Requires
\code{activity:read_all} for Only Me activities.
}
\examples{
if (FALSE) {
  get_activities(7255352296)
}

tidy(strava_data$activities)

}
\concept{activities}
