% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities_comments.R
\name{get_activities_comments}
\alias{get_activities_comments}
\title{List Activity Comments}
\usage{
get_activities_comments(
  id,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the activity.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{An array of [Comment](https://developers.strava.com/docs/reference/#api-models-Comment) objects.
}\if{html}{\out{</div>}}
}
\description{
Returns the comments on the given activity. Requires \code{activity:read} for
Everyone and Followers activities. Requires \code{activity:read_all} for Only Me
activities.
}
\examples{
if (FALSE) {
  get_activities_comments(7148393860)
}

# example response
tidy(strava_data$activities_comments)

}
\concept{activities}
