% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities_kudos.R
\name{get_activities_kudos}
\alias{get_activities_kudos}
\title{List Activity Kudoers}
\usage{
get_activities_kudos(
  id,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the activity.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\value{
An array of \href{https://developers.strava.com/docs/reference/#api-models-SummaryAthlete}{SummaryAthlete} objects.
}
\description{
Returns the athletes who kudoed an activity identified by an identifier.
Requires \code{activity:read} for Everyone and Followers activities. Requires
\code{activity:read_all} for Only Me activities.
}
\examples{
if (FALSE) {
  get_activities_kudos(7255352296)
}

# example response
tidy(strava_data$activities_kudos)

}
\concept{activities}
