% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete.R
\name{get_athlete}
\alias{get_athlete}
\title{Get Authenticated Athlete}
\usage{
get_athlete(scope = "activity:read", client = strava_client())
}
\arguments{
\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the currently authenticated athlete. Tokens with profile:read_all
scope will receive a detailed athlete representation; all others will
receive a summary representation.
}
