% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_activities.R
\name{get_athlete_activities}
\alias{get_athlete_activities}
\title{List Athlete Activities}
\usage{
get_athlete_activities(
  before = NULL,
  after = NULL,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{before}{An epoch timestamp to use for filtering activities that have
taken place before a certain time.}

\item{after}{An epoch timestamp to use for filtering activities that have
taken place after a certain time.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the activities of an athlete for a specific identifier. Requires
activity:read. Only Me activities will be filtered out unless requested by a
token with activity:read_all.
}
\examples{
if (FALSE) {
  get_athlete_activities()
}

}
\concept{athlete}
