% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_clubs.R
\name{get_athlete_clubs}
\alias{get_athlete_clubs}
\title{List Athlete Clubs}
\usage{
get_athlete_clubs(
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a list of the clubs whose membership includes the authenticated
athlete.
}
\concept{athlete}
