% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athlete_stats.R
\name{get_athlete_stats}
\alias{get_athlete_stats}
\title{Get Athlete Stats}
\usage{
get_athlete_stats(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the athlete. Must match the authenticated athlete.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the activity stats of an athlete. Only includes data from activities
set to Everyone visibilty.
}
\concept{athlete}
