% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_athletes_routes.R
\name{get_athletes_routes}
\alias{get_athletes_routes}
\title{List Athlete Routes}
\usage{
get_athletes_routes(
  id,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the athlete. Must match the authenticated athlete.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a list of the routes created by the authenticated athlete. Private
routes are filtered out unless requested by a token with read_all scope.
}
\concept{routes}
