% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clubs.R
\name{get_clubs}
\alias{get_clubs}
\title{Get Club}
\usage{
get_clubs(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the club.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a given club using its identifier.
}
\concept{clubs}
