% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clubs_activities.R
\name{get_clubs_activities}
\alias{get_clubs_activities}
\title{List Club Activities}
\usage{
get_clubs_activities(
  id,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the club.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Retrieve recent activities from members of a specific club. The authenticated
athlete must belong to the requested club in order to hit this endpoint.
Pagination is supported. Athlete profile visibility is respected for all
activities.
}
\concept{clubs}
