% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clubs_members.R
\name{get_clubs_members}
\alias{get_clubs_members}
\title{List Club Members}
\usage{
get_clubs_members(
  id,
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the club.}

\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a list of the athletes who are members of a given club.
}
