% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gear.R
\name{get_gear}
\alias{get_gear}
\title{Get Equipment}
\usage{
get_gear(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the gear.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns an equipment using its identifier.
}
\examples{
if (FALSE) {
  get_gear("b10826293")
}

# example response
tidy(strava_data$gear)

}
