% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes.R
\name{get_routes}
\alias{get_routes}
\title{Get Route}
\usage{
get_routes(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the route.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a route using its identifier. Requires read_all scope for private routes.
}
\examples{
if (FALSE) {
  get_routes("2964569522013822534")
}

# example response
tidy(strava_data$routes)

}
