% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_routes_streams.R
\name{get_routes_streams}
\alias{get_routes_streams}
\title{Get Route Streams}
\usage{
get_routes_streams(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the route.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the given route's streams. Requires read_all scope for private
routes.
}
\examples{
if (FALSE) {
  get_routes_streams("2964569522013822534")
}

}
