% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_efforts.R
\name{get_segment_efforts}
\alias{get_segment_efforts}
\title{List Segment Efforts}
\usage{
get_segment_efforts(
  segment_id,
  start_date_local = NULL,
  end_date_local = NULL,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{segment_id}{The identifier of the segment.}

\item{start_date_local}{ISO 8601 formatted date time.}

\item{end_date_local}{ISO 8601 formatted date time.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a set of the authenticated athlete's segment efforts for a given
segment. Requires subscription.
}
\examples{
if (FALSE) {
  get_segment_efforts(10561790)
}

# example response
tidy(strava_data$segment_efforts)

}
