% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_efforts_by_id.R
\name{get_segment_efforts_by_id}
\alias{get_segment_efforts_by_id}
\title{Get Segment Effort}
\usage{
get_segment_efforts_by_id(
  id,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the segment effort.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns a segment effort from an activity that is owned by the authenticated
athlete. Requires subscription.
}
