% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segment_efforts_streams.R
\name{get_segment_efforts_streams}
\alias{get_segment_efforts_streams}
\title{Get Segment Effort Streams}
\usage{
get_segment_efforts_streams(
  id,
  keys = NULL,
  keys_by_type = TRUE,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the activity.}

\item{keys}{Desired stream types. May take one of the following values:}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}

\item{key_by_type}{Must be true.}
}
\description{
Returns a set of streams for a segment effort completed by the authenticated
athlete. Requires read_all scope.
}
