% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments.R
\name{get_segments}
\alias{get_segments}
\title{Get Segment}
\usage{
get_segments(id, scope = "activity:read", client = strava_client())
}
\arguments{
\item{id}{The identifier of the segment.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the specified segment. read_all scope required in order to retrieve
athlete-specific segment information, or to retrieve private segments.
}
\examples{
if (FALSE) {
  get_segments(1089965)
}

# example response
tidy(strava_data$segments)

}
