% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments_starred.R
\name{get_segments_starred}
\alias{get_segments_starred}
\title{List Starred Segments}
\usage{
get_segments_starred(
  page = 1,
  per_page = 30,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{page}{Page number. Defaults to 1.}

\item{per_page}{Number of items per page. Defaults to 30.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
List of the authenticated athlete's starred segments. Private segments are
filtered out unless requested by a token with read_all scope.
}
