% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zones.R
\name{get_zones}
\alias{get_zones}
\title{Get Zones}
\usage{
get_zones(scope = "profile:read_all", client = strava_client())
}
\arguments{
\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the the authenticated athlete's heart rate and power zones. Requires
profile:read_all.
}
