% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpx_example.R
\name{gpx_example}
\alias{gpx_example}
\title{Get an example gpx}
\usage{
gpx_example(name = NULL)
}
\arguments{
\item{name}{Name of the gpx file. Leave blank to display the list of
examples}
}
\description{
Get an exampe gpx file from the package
}
\examples{
gpx_example("trail_roche_doetre")
}
