% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Display Strava activities as calendar heat map}
\usage{
plot_calendar(data)
}
\arguments{
\item{data}{A data frame output from process_data()}
}
\value{
A plot displaying calendar heat map
}
\description{
Display Strava activities as calendar heat map
}
