% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_facets.R
\name{plot_facets}
\alias{plot_facets}
\title{Creates a plot of activities as small multiples}
\usage{
plot_facets(data, labels = FALSE)
}
\arguments{
\item{data}{A data frame output from process_data()}

\item{labels}{If TRUE, adds distance labels to each facet}
}
\description{
Creates a plot of activities as small multiples. The concept behind this plot was originally inspired by \href{https://www.madewithsisu.com/}{Sisu}.
}
\examples{
plot_facets()
}
