% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_packed_circles.R
\name{plot_packed_circles}
\alias{plot_packed_circles}
\title{Display Strava activities as packed circles}
\usage{
plot_packed_circles(data, circle_size = "duration", circle_fill = "distance")
}
\arguments{
\item{data}{A data frame output from process_data()}

\item{circle_size}{A string ("distance" / "duration" / "speed") defining sizes of circles}

\item{circle_fill}{A string ("distance" / "duration" / "speed") defining fills of circles}
}
\value{
A plot with Strava activities as packed circles
}
\description{
Display Strava activities as packed circles
}
