% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ridges.R
\name{plot_ridges}
\alias{plot_ridges}
\title{Plot ridges of Strava activities by weekdays}
\usage{
plot_ridges(data)
}
\arguments{
\item{data}{A data frame output from process_data()}
}
\value{
A plot displaying ridges
}
\description{
Plot ridges of Strava activities by weekdays
}
