% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Processes gpx files and stores the result in a data frame}
\usage{
process_data(path)
}
\arguments{
\item{path}{The file path to the directory containing the gpx files}
}
\description{
Processes gpx files and stores the result in a data frame
}
\examples{
process_data()
}
