% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava-package.R
\docType{package}
\name{strava-package}
\alias{strava}
\alias{strava-package}
\title{strava: Strava API}
\description{
Easily access data from the 'Strava API' \url{https://developers.strava.com/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tyluRp/strava}
  \item \url{https://tylurp.github.io/strava/}
  \item Report bugs at \url{https://github.com/tyluRp/strava/issues}
}

}
\author{
\strong{Maintainer}: Tyler Littlefield \email{tylerlittlefield@hey.com}

}
\keyword{internal}
