% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_data.R
\docType{data}
\name{strava_data}
\alias{strava_data}
\title{Example strava data}
\format{
An object of class \code{list} of length 10.
}
\usage{
strava_data
}
\description{
A list containing example responses from all \code{get_} functions.
}
\keyword{datasets}
