% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments_explore.R
\name{get_segments_explore}
\alias{get_segments_explore}
\title{Explore segments}
\usage{
get_segments_explore(
  bounds = NULL,
  activity_type = NULL,
  min_cat = NULL,
  max_cat = NULL,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{bounds}{The latitude and longitude for two points describing a
rectangular boundary for the search: (southwest corner latitutde, southwest
corner longitude, northeast corner latitude, northeast corner longitude)}

\item{activity_type}{Desired activity type. May take one of the following
values: \code{running}, \code{riding}}

\item{min_cat}{The minimum climbing category.}

\item{max_cat}{The maximum climbing category.}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}
}
\description{
Returns the top 10 segments matching a specified query.
}
\examples{
if (FALSE) {
  get_segments_explore(
    bounds = "33.8651,-118.8450,34.2065,-118.1886",
    activity_type = "riding"
  )
}

# example response
tidy(strava_data$segments_explore)

}
