% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments_streams.R
\name{get_segments_streams}
\alias{get_segments_streams}
\title{Get Segment Streams}
\usage{
get_segments_streams(
  id,
  keys = NULL,
  keys_by_type = TRUE,
  scope = "activity:read",
  client = strava_client()
)
}
\arguments{
\item{id}{The identifier of the activity.}

\item{keys}{Desired stream types. May take one of the following values:}

\item{scope}{Authorization scope.}

\item{client}{Strava client. By default, assumes you have client credentials
stored in your \code{.Renviron} file.}

\item{key_by_type}{Must be true.}
}
\description{
Returns the given segment's streams. Requires read_all scope for private
segments.
}
