% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3D.R
\name{plot_3D}
\alias{plot_3D}
\title{Create 3D interactive elevation map from gpx tracklogs}
\usage{
plot_3D(
  tracklog_file = "gpx/mtb",
  cache_folder = "~/cache_plot_3D",
  elevation_scale = 6,
  elevation_scale_tracklog_corr = 0.01,
  buffer_around_tracklog_km = 1,
  render_high_quality = FALSE,
  sunangle = 250,
  sunaltitude = 75,
  color_tracklog = "blue",
  color_background = "lightskyblue1",
  water_detect = FALSE
)
}
\arguments{
\item{tracklog_file}{A path to a gpx file to be read by strava::process_data.
If the file covers a big area plot_3D() can slow down significantly.}

\item{cache_folder}{Directory to store the downloaded elevation data and overlay image for reuse}

\item{elevation_scale}{Horizontal vs vertical ratio of the plot. Could be estimated with geoviz::raster_zscale(dem).}

\item{elevation_scale_tracklog_corr}{Raises elevation_scale for the tracklog by the specified percentage.
Useful if tracklog occasionally disappears into the ground.}

\item{buffer_around_tracklog_km}{Buffer distance around the tracklog in km. Scene is cut outside this.}

\item{render_high_quality}{If set to TRUE the function downloads higher resolution elevation and overlay data.
And calculates ambient shadows as well. All steps are time and resource consuming.}

\item{sunangle}{The direction of the light source of the scene}

\item{sunaltitude}{The angle, in degrees (as measured from the horizon) from which the light originates.}

\item{color_tracklog}{Color of the tracklog on the scene.}

\item{color_background}{Color of the scene's background.}

\item{water_detect}{If set to TRUE rayshader will look for flat surfaces in the scene and render them in blue}
}
\description{
Create 3D interactive elevation map from gpx tracklogs
}
\examples{
\dontrun{
# Generate a built in demo by simply running
strava::plot_3D()
}
}
