% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calendar.R
\name{plot_calendar}
\alias{plot_calendar}
\title{Display Strava activities as calendar heat map}
\usage{
plot_calendar(data, unit = "distance")
}
\arguments{
\item{data}{A data frame output from process_data()}

\item{unit}{A string ("distance" or "time") defining which unit to display}
}
\value{
A plot displaying calendar heat map
}
\description{
Display Strava activities as calendar heat map
}
\examples{
running <- process_data(system.file("gpx/running", package = "strava"))
plot_calendar(running)
}
