% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elevations.R
\name{plot_elevations}
\alias{plot_elevations}
\title{Creates a plot of elevation profiles as small multiples}
\usage{
plot_elevations(data)
}
\arguments{
\item{data}{A data frame output from process_data()}
}
\description{
Creates a plot of elevation profiles as small multiples
}
\examples{
plot_elevations()
}
