% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{Plots activities superimposed on a map}
\usage{
plot_map(data, lon_min = 0, lon_max = Inf, lat_min = 0, lat_max = Inf)
}
\arguments{
\item{data}{A data frame output from process_data()}

\item{lon_min}{Minimum longitude}

\item{lon_max}{Maximum longitude}

\item{lat_min}{Minimum latitude}

\item{lat_max}{Maximum latitude}
}
\description{
Plots activities superimposed on a map
}
\examples{
plot_map()
}
