% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{Processes gpx files and stores the result in a data frame}
\usage{
process_data(path, old_gpx_format = FALSE)
}
\arguments{
\item{path}{The file path to the directory containing the gpx files}

\item{old_gpx_format}{If TRUE, uses the old format for gpx files (for files bulk exported from Strava prior to ~May 2018)}
}
\description{
Processes gpx files and stores the result in a data frame. The code is adapted from the blog post \href{https://rcrastinate.blogspot.com/2014/09/stay-on-track-plotting-gps-tracks-with-r.html}{Stay on track: Plotting GPS tracks with R} by Sascha W.
}
