% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strava_client.R
\name{strava_client}
\alias{strava_client}
\title{Create a strava client}
\usage{
strava_client(
  id = Sys.getenv("STRAVA_ID"),
  secret = Sys.getenv("STRAVA_SECRET"),
  token_url = "https://www.strava.com/oauth/token",
  name = "strava-oauth"
)
}
\arguments{
\item{id}{Client ID. By default, attempts to pull ID from .Renviron file.}

\item{secret}{Client secret By default, attempts to pull ID from .Renviron file.}

\item{token_url}{Token URL. Should not need to be changed.}

\item{name}{Client name. By default, 'strava-oauth' but you can name it whatever.}
}
\description{
Create a strava client
}
\examples{
if (FALSE) {
  strava_client()
}

}
