% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activities.R, R/get_gear.R, R/get_routes.R,
%   R/get_segment_efforts.R, R/get_segments.R, R/get_segments_explore.R,
%   R/get_segments_starred.R, R/get_segments_streams.R, R/tidy.R
\name{tidy.strava_activities}
\alias{tidy.strava_activities}
\alias{tidy.strava_gear}
\alias{tidy.strava_routes}
\alias{tidy.strava_segment_efforts}
\alias{tidy.strava_segments}
\alias{tidy.strava_segments_explore}
\alias{tidy.strava_segments_starred}
\alias{tidy.strava_segments_streams}
\alias{tidy}
\title{Tidy a strava response}
\usage{
\method{tidy}{strava_activities}(x, ...)

\method{tidy}{strava_gear}(x, ...)

\method{tidy}{strava_routes}(x, ...)

\method{tidy}{strava_segment_efforts}(x, ...)

\method{tidy}{strava_segments}(x, ...)

\method{tidy}{strava_segments_explore}(x, ...)

\method{tidy}{strava_segments_starred}(x, ...)

\method{tidy}{strava_segments_streams}(x, ...)

tidy(x, ...)
}
\arguments{
\item{x}{A response from any \code{get_} functions.}

\item{...}{Ignore.}
}
\description{
Strava responses are typically returned in the form of a list. This function
will make the response \emph{tidy} by transforming it into a flat table.
}
\concept{helpers}
